test : 인수테스트 url 리팩토링
- 인수테스트의 url을 변경된 url로 리팩토링함.
- reward tag관련 요청을 reward-tag/... 로 변경
