+++ waldreg/application-layer/reward-controller/src/main/java/org/waldreg/controller/reward/tag/RewardController.java
@@ -11,8 +11,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import org.wadlreg.reward.tag.RewardTagManager;
-import org.wadlreg.reward.tag.dto.RewardTagDto;
+import org.waldreg.reward.tag.RewardTagManager;
+import org.waldreg.reward.tag.dto.RewardTagDto;
 import org.waldreg.character.aop.annotation.PermissionVerifying;
 import org.waldreg.controller.reward.tag.mapper.ControllerRewardTagMapper;
 import org.waldreg.controller.reward.tag.request.RewardTagRequest;
