+++ waldreg/application-layer/reward-controller/src/main/java/org/waldreg/controller/reward/tag/mapper/ControllerRewardTagMapper.java
@@ -3,7 +3,7 @@ package org.waldreg.controller.reward.tag.mapper;
 import java.util.ArrayList;
 import java.util.List;
 import org.springframework.stereotype.Service;
-import org.wadlreg.reward.tag.dto.RewardTagDto;
+import org.waldreg.reward.tag.dto.RewardTagDto;
 import org.waldreg.controller.reward.tag.request.RewardTagRequest;
 import org.waldreg.controller.reward.tag.response.RewardTagResponse;
