+++ waldreg/application-layer/reward-controller/src/main/java/org/waldreg/controller/reward/users/mapper/ControllerUsersRewardMapper.java
@@ -3,8 +3,8 @@ package org.waldreg.controller.reward.users.mapper;
 import java.util.ArrayList;
 import java.util.List;
 import org.springframework.stereotype.Service;
-import org.wadlreg.reward.users.dto.UsersRewardDto;
-import org.wadlreg.reward.users.dto.UsersRewardTagDto;
+import org.waldreg.reward.users.dto.UsersRewardDto;
+import org.waldreg.reward.users.dto.UsersRewardTagDto;
 import org.waldreg.controller.reward.users.response.RewardTagWrapperResponse;
 import org.waldreg.controller.reward.users.response.UsersRewardTagResponse;
