+++ waldreg/repository-layer/memory-repository/src/main/java/org/waldreg/repository/reward/RewardTagMapper.java
@@ -1,7 +1,7 @@
 package org.waldreg.repository.reward;
 
 import org.springframework.stereotype.Service;
-import org.wadlreg.reward.tag.dto.RewardTagDto;
+import org.waldreg.reward.tag.dto.RewardTagDto;
 import org.waldreg.domain.rewardtag.RewardTag;
 
 @Service
