+++ waldreg/repository-layer/memory-repository/src/main/java/org/waldreg/repository/usersreward/MemoryUsersRewardRepository.java
@@ -3,9 +3,9 @@ package org.waldreg.repository.usersreward;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
-import org.wadlreg.reward.users.dto.UsersRewardDto;
-import org.wadlreg.reward.users.spi.repository.UserExistChecker;
-import org.wadlreg.reward.users.spi.repository.UsersRewardRepository;
+import org.waldreg.reward.users.dto.UsersRewardDto;
+import org.waldreg.reward.users.spi.repository.UserExistChecker;
+import org.waldreg.reward.users.spi.repository.UsersRewardRepository;
 import org.waldreg.domain.rewardtag.RewardTag;
 import org.waldreg.domain.rewardtag.RewardTagWrapper;
 import org.waldreg.domain.user.User;
