+++ waldreg/repository-layer/memory-repository/src/test/java/org/waldreg/repository/usersreward/MemoryUsersRewardRepositoryTest.java
@@ -11,8 +11,8 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.wadlreg.reward.users.dto.UsersRewardDto;
-import org.wadlreg.reward.users.dto.UsersRewardTagDto;
+import org.waldreg.reward.users.dto.UsersRewardDto;
+import org.waldreg.reward.users.dto.UsersRewardTagDto;
 import org.waldreg.domain.character.Character;
 import org.waldreg.domain.rewardtag.RewardTag;
 import org.waldreg.domain.user.User;
