+++ waldreg/service-layer/reward/src/main/java/org/waldreg/reward/config/RewardPermissionConfigurer.java
@@ -1,4 +1,4 @@
-package org.wadlreg.reward.config;
+package org.waldreg.reward.config;
 
 import java.util.List;
 import javax.annotation.PostConstruct;
