+++ waldreg/service-layer/reward/src/main/java/org/waldreg/reward/tag/dto/RewardTagDto.java
@@ -1,4 +1,4 @@
-package org.wadlreg.reward.tag.dto;
+package org.waldreg.reward.tag.dto;
 
 public final class RewardTagDto{
