+++ waldreg/service-layer/reward/src/main/java/org/waldreg/reward/tag/spi/RewardTagRepository.java
@@ -1,9 +1,10 @@
-package org.wadlreg.reward.tag;
+package org.waldreg.reward.tag.spi;
 
 import java.util.List;
-import org.wadlreg.reward.tag.dto.RewardTagDto;
+import org.waldreg.reward.tag.dto.RewardTagDto;
+import org.waldreg.reward.users.spi.tag.RewardTagExistChecker;
 
-public interface RewardTagManager{
+public interface RewardTagRepository extends RewardTagExistChecker{
 
     void createRewardTag(RewardTagDto rewardTagDto);
