+++ waldreg/service-layer/reward/src/test/java/org/waldreg/reward/tag/RewardTagManagerTest.java
@@ -10,12 +10,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.wadlreg.reward.tag.DefaultRewardTagManager;
-import org.wadlreg.reward.tag.RewardTagManager;
-import org.wadlreg.reward.tag.dto.RewardTagDto;
-import org.wadlreg.reward.exception.UnknownRewardTagException;
-import org.wadlreg.reward.tag.lib.TagExceedClipper;
-import org.wadlreg.reward.tag.spi.RewardTagRepository;
+import org.waldreg.reward.tag.dto.RewardTagDto;
+import org.waldreg.reward.exception.UnknownRewardTagException;
+import org.waldreg.reward.tag.lib.TagExceedClipper;
+import org.waldreg.reward.tag.spi.RewardTagRepository;
 
 @ExtendWith(SpringExtension.class)
 @ContextConfiguration(classes = {DefaultRewardTagManager.class, TagExceedClipper.class})
