+++ waldreg/service-layer/reward/src/test/java/org/waldreg/reward/users/UsersRewardManagerTest.java
@@ -10,16 +10,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.wadlreg.reward.exception.UnknownRewardException;
-import org.wadlreg.reward.exception.UnknownRewardTargetException;
-import org.wadlreg.reward.exception.UnknownRewardTagException;
-import org.wadlreg.reward.users.DefaultUsersRewardManager;
-import org.wadlreg.reward.users.UsersRewardManager;
-import org.wadlreg.reward.users.dto.UsersRewardDto;
-import org.wadlreg.reward.users.dto.UsersRewardTagDto;
-import org.wadlreg.reward.users.spi.repository.UserExistChecker;
-import org.wadlreg.reward.users.spi.repository.UsersRewardRepository;
-import org.wadlreg.reward.users.spi.tag.RewardTagExistChecker;
+import org.waldreg.reward.exception.UnknownRewardException;
+import org.waldreg.reward.exception.UnknownRewardTargetException;
+import org.waldreg.reward.exception.UnknownRewardTagException;
+import org.waldreg.reward.users.dto.UsersRewardDto;
+import org.waldreg.reward.users.dto.UsersRewardTagDto;
+import org.waldreg.reward.users.spi.repository.UserExistChecker;
+import org.waldreg.reward.users.spi.repository.UsersRewardRepository;
+import org.waldreg.reward.users.spi.tag.RewardTagExistChecker;
 
 @ExtendWith(SpringExtension.class)
 @ContextConfiguration(classes = DefaultUsersRewardManager.class)
